requirements.txt
matplotlib==3.5.3
modbus_tk==1.1.2
numpy==1.23.2
pandas==1.4.3
Pillow==9.2.0
pyserial==3.5
python-dotenv==0.21.0
PyVISA==1.11.3
scipy==1.9.0
setuptools==65.3.0
slackclient==2.9.4
